add numOfDiner for Dish

+++ source/ServerSource/liferay-plugins-sdk-6.1.1/portlets/DishStore-portlet/docroot/WEB-INF/service/irestads/service/DishLocalServiceClp.java
@@ -108,7 +108,7 @@ public class DishLocalServiceClp implements DishLocalService {
 
 		_methodParameterTypes19 = new String[] {
 				"java.lang.String", "java.lang.String", "java.lang.String",
-				"java.lang.String", "java.lang.String", "double", "long"
+				"java.lang.String", "java.lang.String", "int", "long"
 			};
 
 		_methodName20 = "getAllDishs";
@@ -640,7 +640,7 @@ public class DishLocalServiceClp implements DishLocalService {
 
 	public irestads.model.Dish createDish(java.lang.String dishName,
 		java.lang.String decription, java.lang.String avatarImg,
-		java.lang.String detailImg, java.lang.String detail, double referPrice,
+		java.lang.String detailImg, java.lang.String detail, int referPrice,
 		long categoryId) {
 		Object returnObj = null;
