+++ source/ServerSource/liferay-plugins-sdk-6.1.1/portlets/DishStore-portlet/docroot/WEB-INF/service/irestads/service/DishLocalServiceWrapper.java
@@ -260,7 +260,7 @@ public class DishLocalServiceWrapper implements DishLocalService,
 
 	public irestads.model.Dish createDish(java.lang.String dishName,
 		java.lang.String decription, java.lang.String avatarImg,
-		java.lang.String detailImg, java.lang.String detail, double referPrice,
+		java.lang.String detailImg, java.lang.String detail, int referPrice,
 		long categoryId) {
 		return _dishLocalService.createDish(dishName, decription, avatarImg,
 			detailImg, detail, referPrice, categoryId);
